requirements.txt
absl-py==0.12.0
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.36.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.20.1
oauthlib==3.1.0
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.15.0
sqlparse==0.4.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.26.4
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
